Add wipe-content Module. IN-PROGESS**
In terms of how the wipe feature is written, im pretty certain that its correctly written to function how I want. The issue I am having is that when aclling this module for a button click event, it runs before the button is clicked.  This results in a blank page.