Add Leading Zero to Days < 10
When the user would not enter a due date, the program defaulted to todays date. All parts of the date has a minumum of 2 digits, but it wasnting adding a leading 0 if the number was less than 10. This has been fixed by converting the day into a string and addinga '0' in front of it