Add A Nearly Complete Way to Project Display Tasks
        build(todoListRows(iterator.title, iterator.dueDate, iterator.status, iterator.id).row)
conceptually, all the pieces I need to show the the project tasks are in this updateProjectTasksLS() Method. The only problem is that I am invoking todoListRows()  to create the rows. The problem with this is that this is the the widths of each row is based on the current  adjustable width of the task headers on the different webpages. I may have to create a seprate module that creates a row for projects with mostly duplicate code (just to save time), so that the rows will properly fit in the viewpoint where the tasks are shown